fix: preserve modern temporal defaults
